remove separate version file
