--- cli/src/version.json
@@ -1,3 +1,0 @@
-{
-  "version": "0.0.4"
-}
